refactor(srv): split service/workspace.rs and window.rs match arms into per-domain files
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Each file was one giant match over call.method with 16-18 fully independent
async handler arms and no shared private state. workspace.rs (1358 lines) ->
workspace_lifecycle/tab_lifecycle/tab_move/tear_off.rs + a ~80-line dispatcher.
window.rs (1327 lines) -> window_query/window_create/window_close/
window_mutate.rs (each carrying its already-existing feature-scoped test
module) + a ~40-line dispatcher.

Pure extraction, no logic changes, handler signatures unchanged
(handle_workspace_service/handle_window_service entry points untouched, so
no external call sites needed changes). cargo build -p agentmux-srv clean;
cargo test -p agentmux-srv service:: — 47 passed, 0 failed, same test counts
as before the split.

<!-- agentmux:agent_id=agentx -->
